# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API ",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/"
        }
    ],
    "paths": {
        "/api/AllUsers": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "AllUsers",
                "operationId": "AllUsers",
                "responses": {
                    "200": {
                        "description": "All USers data",
                        "content": {
                            "application/json": {}
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/SinglePageUser": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "SinglePageUser",
                "operationId": "SinglePageUser",
                "parameters": [
                    {
                        "name": "user_id",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User Data And relation Car",
                        "content": {
                            "application/json": {}
                        }
                    }
                }
            }
        },
        "/api/CreateCar": {
            "post": {
                "tags": [
                    "Car"
                ],
                "summary": "CreateCar",
                "operationId": "CreateCar",
                "parameters": [
                    {
                        "name": "model",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "date",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "price",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "photo[ ]",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "file"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User Login end",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "422": {
                        "description": "Validation Error"
                    },
                    "423": {
                        "description": "Wrong Email or password"
                    },
                    "424": {
                        "description": "You have a car"
                    }
                }
            }
        },
        "/api/DeleteCar": {
            "get": {
                "tags": [
                    "Car"
                ],
                "summary": "DeleteCar",
                "operationId": "DeleteCar",
                "parameters": [
                    {
                        "name": "car_id",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Car Deleted",
                        "content": {
                            "application/json": {}
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/register": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "register",
                "operationId": "register",
                "parameters": [
                    {
                        "name": "name",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "email",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "password",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "password_confirmation",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User Created",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "422": {
                        "description": "Validation Error"
                    }
                }
            }
        },
        "/api/login": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "login",
                "operationId": "login",
                "parameters": [
                    {
                        "name": "email",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "password",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User Login end",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "422": {
                        "description": "Validation Error"
                    },
                    "423": {
                        "description": "Wrong Email or password"
                    }
                }
            }
        },
        "/api/logout": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "logout",
                "operationId": "logout",
                "responses": {
                    "200": {
                        "description": "User Logout end",
                        "content": {
                            "application/json": {}
                        }
                    }
                },
                "security": []
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Login with email and password to get the authentication token",
                "name": "Token based Based",
                "in": "header",
                "bearerFormat": "bearerAuth",
                "scheme": "bearerAuth"
            }
        }
    }
}